% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_factors_tidy.R
\name{prepare_factors_tidy}
\alias{prepare_factors_tidy}
\title{Preparing data for machine learning}
\usage{
prepare_factors_tidy(train, test = NULL, rare_count = 5L,
  rare_level = "rare", new_level = "other", ignore_na = TRUE,
  output_type = c("character", "factor", "integer"), encoding = NULL,
  return_mapping = FALSE)
}
\arguments{
\item{train}{Character. Possibly a feature column from a training data set.}

\item{test}{Character. Possibly a feature column from a test data set.}

\item{rare_count}{Integer. Levels with less manifestations than \code{rare_count} will get the \code{rare_level}.}

\item{rare_level}{Character. Specification of the new name for rare levels.}

\item{new_level}{Character. The level for manifestations that didn't show up in the training data.}

\item{ignore_na}{Logical. Should \code{NA}s \code{NA}?}

\item{output_type}{Character ("character", "factor" or "integer"). If \code{"factor"} the output always gets the levels \code{rare_level} and \code{other_level}.}

\item{encoding}{Character. Specifies the sorting of the levels / integers (when the \code{output_type} is not \code{"character"}.}

\item{return_mapping}{Logical. Should a mapping between integers and original characters be returned.}
}
\value{
A contatenated vector of `train` and `test` transformed and converted regarding the \code{output_type}.

### preparation
prepare_factors(train$some_new, test$some_new, output_type = "factor")
}
\description{
Function to convert rare and new levels in test data
}
